requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
asyncpg==0.31.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.1
filetype==1.2.0
frozenlist==1.8.0
google-ai-generativelanguage==0.10.0
google-api-core==2.30.0
google-auth==2.49.1
googleapis-common-protos==1.73.0
greenlet==3.2.5
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.28
langchain-community==0.3.31
langchain-core==0.3.83
langchain-google-genai==2.1.12
langchain-openai==0.3.35
langchain-postgres==0.0.17
langchain-text-splitters==0.3.11
langsmith==0.4.37
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.0.2
openai==2.28.0
orjson==3.11.5
packaging==25.0
pgvector==0.3.6
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg==3.2.13
psycopg-binary==3.2.13
psycopg-pool==3.2.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pypdf==6.8.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
yarl==1.22.0
zstandard==0.25.0
